requirements.txt
Django==4.1.5
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
Pillow==9.4.0

pytest==7.2.1
pytest-django==4.5.2
pytest-mock==3.10.0
factory-boy==3.2.1